+++ packages/remote-logs/CHANGELOG.md
@@ -1,5 +1,111 @@
 # Change Log
 
+## 3.0.16
+
+### Patch Changes
+
+- Updated dependencies [282bb4e]
+  - @bifold/core@3.0.16
+
+## 3.0.15
+
+### Patch Changes
+
+- Updated dependencies [9ed3b0f]
+- Updated dependencies [fc4eeb1]
+- Updated dependencies [7fce466]
+  - @bifold/core@3.0.15
+
+## 3.0.14
+
+### Patch Changes
+
+- Updated dependencies [c70df9d]
+- Updated dependencies [4704ce3]
+  - @bifold/core@3.0.14
+
+## 3.0.13
+
+### Patch Changes
+
+- Updated dependencies [f813512]
+  - @bifold/core@3.0.13
+
+## 3.0.12
+
+### Patch Changes
+
+- Updated dependencies [dd22c43]
+  - @bifold/core@3.0.12
+
+## 3.0.11
+
+### Patch Changes
+
+- 178d2f8: logging change had strange side effects
+  - @bifold/core@3.0.11
+
+## 3.0.10
+
+### Patch Changes
+
+- 149265d: added a custom log replacer to handle nested error objects
+- Updated dependencies [e3d82d0]
+  - @bifold/core@3.0.10
+
+## 3.0.9
+
+### Patch Changes
+
+- Updated dependencies [0944c44]
+  - @bifold/core@3.0.9
+
+## 3.0.8
+
+### Patch Changes
+
+- Updated dependencies [75e0535]
+  - @bifold/core@3.0.8
+
+## 3.0.7
+
+### Patch Changes
+
+- Updated dependencies [7cb7ff0]
+  - @bifold/core@3.0.7
+
+## 3.0.6
+
+### Patch Changes
+
+- Updated dependencies [8d27002]
+  - @bifold/core@3.0.6
+
+## 3.0.5
+
+### Patch Changes
+
+- Updated dependencies [4be9c48]
+- Updated dependencies [cc8ce46]
+- Updated dependencies [21d6689]
+- Updated dependencies [becac32]
+  - @bifold/core@3.0.5
+
+## 3.0.4
+
+### Patch Changes
+
+- Updated dependencies [d0f1003]
+  - @bifold/core@3.0.4
+
+## 3.0.3
+
+### Patch Changes
+
+- Updated dependencies [25146c1]
+- Updated dependencies [c7b866a]
+  - @bifold/core@3.0.3
+
 ## 3.0.2
 
 ### Patch Changes